language-roleplayer/requirements.txt
# === Saynario — Backend Dependencies ===

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9

# AI / ML services
anthropic>=0.40.0
openai>=1.30.0
elevenlabs>=1.0.0

# Audio processing
pydub>=0.25.1
numpy>=1.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Redis
redis>=5.0.0

# Structured output
instructor>=1.0.0

# Config and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio>=4.0.0
httpx[socks]>=0.27.0
websockets>=12.0

# ── Heavy ML (optional — only needed for local Whisper/VAD; not required in MOCK_MODE) ──
# Uncomment these if you want to run Whisper and Silero VAD locally (requires GPU):
# torch>=2.0.0
# torchaudio>=2.0.0
# silero-vad>=4.0.0
# openai-whisper
# language-tool-python>=2.7.0
